requirements.txt
deepmerge>=0.3
click>=8.0
pyyaml>=5.4
requests>=2.26
git-remote-codecommit>=1.15
boto3>=1.18